pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ecommerce-data-platform"
version = "0.1.0"
description = "E-commerce real-time data pipeline platform on GCP/BigQuery with Kubernetes orchestration and GenAI-powered observability"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Jimin"},
]

dependencies = [
    # GCP / BigQuery
    "google-cloud-bigquery>=3.25.0",
    "google-cloud-pubsub>=2.23.0",
    "google-cloud-storage>=2.18.0",
    "google-cloud-logging>=3.11.0",

    # Data Processing
    "apache-beam[gcp]>=2.58.0",
    "pandas>=2.2.0",
    "pyarrow>=17.0.0",

    # Web Framework (Event Collector)
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "pydantic>=2.9.0",

    # GenAI
    "openai>=1.50.0",
    "google-cloud-aiplatform>=1.70.0",
    "langchain>=0.3.0",
    "langchain-google-genai>=2.0.0",

    # Observability
    "prometheus-client>=0.21.0",
    "opentelemetry-api>=1.27.0",
    "opentelemetry-sdk>=1.27.0",
    "opentelemetry-exporter-gcp-trace>=1.8.0",
    "structlog>=24.4.0",

    # Database / CDC
    "sqlalchemy>=2.0.35",
    "psycopg2-binary>=2.9.9",

    # Utilities
    "pyyaml>=6.0.2",
    "click>=8.1.7",
    "python-dotenv>=1.0.1",
    "httpx>=0.27.0",
    "croniter>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.7.0",
    "mypy>=1.12.0",
    "pre-commit>=4.0.0",
    "faker>=30.0.0",
]

airflow = [
    "apache-airflow>=2.10.0",
    "apache-airflow-providers-google>=10.23.0",
]

[project.scripts]
event-collector = "src.event_collector.app:main"
pipeline-cli = "src.cli:main"

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --cov=src --cov-report=term-missing"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
